Update JoystickController.h
Added == and != operators for all the struct definitions in the JoystickController class.  Also added functions eq() and ne() that compares a subset of the State struct.  The intention is to enable a comparison between the current state and previous state of a joystick so that code may be executed only when there is a change in the state.